+++ apps/api/v2/src/modules/oauth/flow/oauth-flow.module.ts
@@ -1,6 +1,9 @@
 import { getEnv } from "@/env";
 import { OAuthFlowController } from "@/modules/oauth/flow/oauth-flow.controller";
 import { OAuthFlowService } from "@/modules/oauth/flow/oauth-flow.service";
+import { OAuthClientModule } from "@/modules/oauth/oauth-client.module";
+import { PrismaModule } from "@/modules/prisma/prisma.module";
+import { TokensRepository } from "@/modules/tokens/tokens.repository";
 import { Module } from "@nestjs/common";
 import { JwtModule } from "@nestjs/jwt";
 
@@ -9,9 +12,11 @@ import { JwtModule } from "@nestjs/jwt";
     JwtModule.register({
       secret: getEnv("NEXTAUTH_SECRET"),
     }),
+    PrismaModule,
+    OAuthClientModule,
   ],
   controllers: [OAuthFlowController],
   exports: [OAuthFlowService],
-  providers: [OAuthFlowService],
+  providers: [OAuthFlowService, TokensRepository],
 })
 export class OAuthFlowModule {}
