+++ apps/api/v2/src/modules/oauth/oauth-client.module.ts
@@ -1,7 +1,6 @@
 import { getEnv } from "@/env";
 import { AuthModule } from "@/modules/auth/auth.module";
 import { MembershipModule } from "@/modules/membership/membership.module";
-import { OAuthFlowModule } from "@/modules/oauth/flow/oauth-flow.module";
 import { OAuthClientGuard } from "@/modules/oauth/guard/oauth-client/oauth-client.guard";
 import { OAuthClientController } from "@/modules/oauth/oauth-client.controller";
 import { OAuthClientRepository } from "@/modules/oauth/oauth-client.repository";
@@ -18,7 +17,6 @@ import { JwtModule } from "@nestjs/jwt";
     UserModule,
     MembershipModule,
     JwtModule.register({ secret: getEnv("JWT_SECRET") }),
-    OAuthFlowModule,
   ],
   providers: [OAuthClientRepository, OAuthClientGuard],
   controllers: [OAuthClientController],
